Add list of more recent changes
This is not a comprehensive list, by far.  But it is a list of the 'biggest hits' and major changes to the code. Once this is added, it should be easier to maintain, in the future.  Though, ideally, we'd still want to figure out how to automate this via Travis CI.